style(scripts): apply the ruff lint and format baseline
Mechanical pass to make `ruff check` and `ruff format --check` clean:
import sorting, pyupgrade rewrites, comprehension and simplification fixes,
f-strings without placeholders, unused imports, trailing whitespace and
consistent quoting.

Substantive items included here:

- B904: preserve the original exception context when re-raising in the GB
  boot ROM assembler.
- F841: drop a dead `opcode` local in assemble_boot_rom.py.
- Mutable class attributes on the Textual screens are now annotated ClassVar.
- Removes a stale `# pylint: disable` in test_sort_roms.py; pylint is not a
  dependency, is not configured and never runs.

Refs #2847

+++ scripts/display_audio_output.py
@@ -29,7 +29,7 @@ def _clamp(value: float, min_value: float, max_value: float) -> float:
 
 def _float_to_u8(sample: float) -> int:
     """Convert a normalized sample in [-1, 1] to unsigned 8-bit [0, 255]."""
-    return int(round((_clamp(sample, -1.0, 1.0) + 1.0) * 127.5))
+    return round((_clamp(sample, -1.0, 1.0) + 1.0) * 127.5)
 
 
 def _read_wav_samples(path: Path) -> list[int]:
@@ -110,9 +110,7 @@ def parse_args() -> argparse.Namespace:
         --start: 0
         --normalize: False
     """
-    parser = argparse.ArgumentParser(
-        description="Plot raw 8-bit audio samples from a .raw or .wav file."
-    )
+    parser = argparse.ArgumentParser(description="Plot raw 8-bit audio samples from a .raw or .wav file.")
     parser.add_argument(
         "path",
         nargs="?",
@@ -151,7 +149,7 @@ def main() -> None:
         raise SystemExit("Start index out of range")
 
     end = min(len(samples), args.start + args.max_samples)
-    window = samples[args.start:end]
+    window = samples[args.start : end]
 
     if args.normalize:
         values = normalize_unsigned_8bit(window)
